+++ tests/src/test_DataSetXsd.cpp
@@ -96,7 +96,6 @@ TEST(DataSetXsdTest, EditDatasetRegistry)
     dataset.Tags("barcode moreTags mapping mytags");
     dataset.TimeStampedName("my_time_stamped_name");
     dataset.UniqueId("b095d0a3-94b8-4918-b3af-a3f81bbe519c");
-    dataset.Version("2.3.0");
     dataset.Attribute("xmlns",              "http://pacificbiosciences.com/PacBioDatasets.xsd")
            .Attribute("xmlns:xsi",          "http://www.w3.org/2001/XMLSchema-instance")
            .Attribute("xsi:schemaLocation", "http://pacificbiosciences.com/PacBioDatasets.xsd");
@@ -113,18 +112,71 @@ TEST(DataSetXsdTest, EditDatasetRegistry)
         "<pbds:AlignmentSet CreatedAt=\"2015-01-27T09:00:01\" MetaType=\"PacBio.DataSet.AlignmentSet\" "
                 "Name=\"DataSet_AlignmentSet\" Tags=\"barcode moreTags mapping mytags\" "
                 "TimeStampedName=\"my_time_stamped_name\" "
-                "UniqueId=\"b095d0a3-94b8-4918-b3af-a3f81bbe519c\" Version=\"2.3.0\" "
+                "UniqueId=\"b095d0a3-94b8-4918-b3af-a3f81bbe519c\" Version=\"3.0.1\" "
                 "xmlns=\"http://pacificbiosciences.com/PacBioDatasets.xsd\" "
                 "xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" "
                 "xsi:schemaLocation=\"http://pacificbiosciences.com/PacBioDatasets.xsd\" "
                 "xmlns:custom=\"http://custom/uri.xsd\" "
                 "xmlns:pbds=\"http://pacificbiosciences.com/PacBioDatasets.xsd\">\n"
         "\t<custom:ExternalResources>\n"
-        "\t\t<custom:ExternalResource MetaType=\"Fake.MetaType\" ResourceId=\"filename\" TimeStampedName=\"custom_tsn\" UniqueId=\"my_uuid\" />\n"
+        "\t\t<custom:ExternalResource MetaType=\"Fake.MetaType\" ResourceId=\"filename\" TimeStampedName=\"custom_tsn\" UniqueId=\"my_uuid\" Version=\"3.0.1\" />\n"
         "\t</custom:ExternalResources>\n"
         "</pbds:AlignmentSet>\n";
 
     stringstream s;
     dataset.SaveToStream(s);
     EXPECT_EQ(expectedXml, s.str());
+}
+
+TEST(DataSetXsdTest, ElementRegistryOk)
+{
+    { // default namespaces
+
+        DataSet ds;
+
+        // append child elements that do not have a C++ built-in, nor namespace prefix with addition
+        DataSetMetadata& metadata = ds.Metadata();
+        metadata.AddChild(internal::DataSetElement("SummaryStats"));
+        metadata.AddChild(internal::DataSetElement("CopyFiles"));
+        metadata.AddChild(internal::DataSetElement("BioSamples"));
+        metadata.AddChild(internal::DataSetElement("AutomationParameters"));
+
+        stringstream s;
+        ds.SaveToStream(s);
+        const string output = s.str();
+
+        // check that default namespace is propagated properly
+        EXPECT_TRUE(output.find("pbds:SummaryStats") != string::npos);
+        EXPECT_TRUE(output.find("pbmeta:CopyFiles") != string::npos);
+        EXPECT_TRUE(output.find("pbsample:BioSamples") != string::npos);
+        EXPECT_TRUE(output.find("pbbase:AutomationParameters") != string::npos);
+    }
+
+    { // custom namespaces
+
+        DataSet ds;
+
+        // setup custom namespaces
+        ds.Namespaces().Register(XsdType::BASE_DATA_MODEL,     NamespaceInfo("custom_base",   "http://custom/base.xsd"));
+        ds.Namespaces().Register(XsdType::COLLECTION_METADATA, NamespaceInfo("custom_meta",   "http://custom/meta.xsd"));
+        ds.Namespaces().Register(XsdType::DATASETS,            NamespaceInfo("custom_ds",     "http://custom/datasets.xsd"));
+        ds.Namespaces().Register(XsdType::SAMPLE_INFO,         NamespaceInfo("custom_sample", "http://custom/base.xsd"));
+
+        // append child elements that do not have a C++ built-in, nor namespace prefix with addition
+        DataSetMetadata& metadata = ds.Metadata();
+        metadata.AddChild(internal::DataSetElement("SummaryStats"));
+        metadata.AddChild(internal::DataSetElement("CopyFiles"));
+        metadata.AddChild(internal::DataSetElement("BioSamples"));
+        metadata.AddChild(internal::DataSetElement("AutomationParameters"));
+
+        stringstream s;
+        ds.SaveToStream(s);
+        const string output = s.str();
+
+        // check that custom namespace is propagated properly
+        EXPECT_TRUE(output.find("custom_ds:SummaryStats") != string::npos);
+        EXPECT_TRUE(output.find("custom_meta:CopyFiles") != string::npos);
+        EXPECT_TRUE(output.find("custom_sample:BioSamples") != string::npos);
+        EXPECT_TRUE(output.find("custom_base:AutomationParameters") != string::npos);
+    }
 }