+++ CHANGELOG.md
@@ -10,6 +10,11 @@ series.
 
 ## Active
 
+## [0.2.3] - 2015-10-26
+- Syncing DataSetXML across APIs. Primary changes include output of Version attribute ("3.0.1") on 
+appropriate elements, as well as resolution of namespace issues.
+   
+
 ## [0.2.2] - 2015-10-22
 
 ### Added